==> 00.model_fits_data_exploration.Rmd <==
---
title: "Untitled"
author: "cmm"
date: '2022-06-27'
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(gratia)
library(tidyverse)
library(mgcv)
library(gridExtra)
source('12.gam_fns.R')

hab_loc_codes <- read_csv('data/habtypes.csv')
hab_loc_codes <- hab_loc_codes %>% select(SETTING_ID, habclass, locationcode)
bai.train <- readRDS('data/bai.train.7_14_22.rds')
bai.train <- left_join(bai.train, hab_loc_codes, by = "SETTING_ID")

bai.spmx <- bai.train %>% 
  select(SETTING_ID, stand, MEASUREMENT_NO, myear, PLOT, cluster, unique_tree_id, 
         bai, DIAMETER, treeba, log.diam, log.bai, grwth.yr, mean_si, aspect_deg, 
         heatload, slope_deg, elev_m, NF, CROWN_RATIO, tpa.pl.all, tpa.pl.cutoff,
         ba.pl, bal.pl, ccf.pl, ccf.nospp, qmd.pl.all, qmd.pl.cutoff, dq.pl.all, 
         dq.pl.cutoff, tpa.cl.all, tpa.pl.cutoff,
         ba.cl, ccf.cl, bal.cl, dq.cl.all, dq.cl.cutoff, HabType, habclass,
         locationcode, percent_LAOC, shade.tol.pl, dom.spp.pl.ba, percent_PSME, 
         percent_PIEN, percent_ABLA, percent_ABGR, percent_PICO, percent_PIPO, percent_other) %>% 
  mutate(slope_pct = tan(slope_deg*(pi/180))*100, 
         asp_sin = sin(aspect_deg*(pi/180)),
         asp_cos = cos(aspect_deg*(pi/180))) %>%
  mutate(sl_asp_sin = asp_sin*slope_pct, 
         sl_asp_cos = asp_cos*slope_pct, 
         asp.trasp = trasp(aspect_deg), 
         HabType = as.factor(HabType),
         cr = CROWN_RATIO/100,
         bal.pl.ratio = bal.pl/ba.pl,
         unique_tree_id = factor(unique_tree_id)) %>% 
  filter(bai != 0)


bai.spmx.ids <- bai.spmx %>% 
  group_by(stand) %>% 
  mutate(unique.stand = cur_group_id()) %>%
  ungroup() %>% 
  mutate(unique.cluster = (1000*unique.stand)+cluster, 
         unique.plot = (1000*unique.stand)+PLOT) %>% 
  mutate(unique.cluster.meas = as.factor(unique.cluster + MEASUREMENT_NO), 
         unique.plot.meas = as.factor(unique.plot + MEASUREMENT_NO),
         unique.cluster.f = as.factor(unique.cluster), 
         unique.plot.f = as.factor(unique.plot), 
         unique.tree.f = as.factor(unique_tree_id)) %>% 
  filter(unique_tree_id != 15431)

#null model to compare to for results
#re.tree.1alt <- readRDS('data/model_objects.1/re.tree.1alt.rds')
#larch fraction - alternative with random 
spmx.gam.alt.lf <- readRDS('data/model_objects.1/spmx.gam.alt.lf_7.11.rds')
#shade tol - alternative with random
spmx.gam.alt.st <- readRDS('data/model_objects.1/spmx.gam.alt.st.rds')

#ccf_noslope <- readRDS('data/model_objects.1/ccf_noslope.rds')

#ccf model
# ccf.gam <- gamm(bai~s(DIAMETER) + s(cr, k = 9) + s(bal.pl.ratio) + s(asp_sin, asp_cos) + s(ccf.pl), random = list(unique_tree_id=~1) , family = 'Gamma'(link = log), data = bai.spmx.ids, method = 'ML', control = list(nthreads = 3))
# ccf.gam1 <- gamm4::gamm4(bai~s(DIAMETER) + s(cr, k = 9) + s(bal.pl.ratio) + s(asp_sin, asp_cos) + s(ccf.pl), random=~(1|unique_tree_id) , family = 'Gamma'(link = log), data = bai.spmx.ids, method = 'ML')
# 
# summary(ccf.gam$gam)
# ccf.gam$lme
# 
# ff <- exp(fitted(ccf.gam$lme))

# a <- summary(re.tree.1alt)
# aa <- summary(ccf_noslope)
# b <- summary(spmx.gam.alt.lf)
# c <- summary(spmx.gam.alt.st)
```

```{r}
#### Here is where I fit models which look at a weighted density metric (BASAL AREA). Comparing larch % weighted ba, shade intolerancce weighted ba, and crown competition factor. 

### Interestingly, the functional relationship among all of them is very similar.

## Curious what I will see after accounting for random effets
## curious what it would look like if instead these were accounted for through the addition of a interaction term. 
## This may mean that crown competition factor is perfectly adequately accounting for species mixing.. HOWEVER.. the ba term also functionally looks the same, which might suggest that CCF is mostly accounting for density..(which is what we already knew..)
## how do i figure out how much of species mixture is accounted for by ccf?

#### Next step is to parse this out and add random effects. Also evaluate 

bai.shade <- bai.spmx.ids %>% 
  mutate(shade.tol.ba = ba.pl*shade.tol.pl, larch.ba = larch.ba.fraction.pl*ba.pl)

ba.shade <- gam(bai~s(DIAMETER) + s(cr, k = 9) + s(bal.pl.ratio) + s(shade.tol.ba) + s(asp_sin, asp_cos), family = 'Gamma'(link = log), data = bai.shade, method = 'ML', control = list(nthreads = 3))
ba.lf <- gam(bai~s(DIAMETER) + s(cr, k = 9) + s(bal.pl.ratio) + s(larch.ba) + s(asp_sin, asp_cos), family = 'Gamma'(link = log), data = bai.shade, method = 'ML', control = list(nthreads = 3))
ccf.abc <- gam(bai~s(DIAMETER) + s(cr, k = 9) + s(bal.pl.ratio) + s(ccf.pl) + s(asp_sin, asp_cos), family = 'Gamma'(link = log), data = bai.shade, method = 'ML', control = list(nthreads = 3))

summary(ba.shade)
plot(ba.shade)

summary(ba.lf)
plot(ba.lf)

summary(ccf.abc)
plot(ccf.abc)

plot(re.tree.1alt)
# saveRDS(spmx.gam.lf, 'data/model_objects.1/spmx.gam.lf.rds')
```


```{r}
#compare spp mix models to null model
my.rmse.2(re.tree.1alt)[[2]]
my.rmse.2(spmx.gam.alt.lf)[[2]]
my.rmse.2(spmx.gam.alt.st)[[2]]

a$dev.expl
b$dev.expl
c$dev.expl
```

```{r}
#fit nospp ccf and spp-informed ccf models
bai.nospp <- bai.spmx.ids %>% mutate(ccf.pl = ccf.nospp)

ccf.model <- gam(bai~s(DIAMETER) + s(cr, k = 9) + s(bal.pl.ratio) + s(ccf.pl) + s(asp_sin, asp_cos), family = 'Gamma'(link = log), data = bai.spmx.ids, method = 'ML', control = list(nthreads = 3))

ccf.nospp <- gam(bai~s(DIAMETER) + s(cr, k = 9) + s(bal.pl.ratio) + s(ccf.pl) + s(asp_sin, asp_cos), family = 'Gamma'(link = log), data = bai.nospp, method = 'ML', control = list(nthreads = 3))

plot.gam(ccf.model, rug = F)
plot.gam(ccf.nospp, rug = F)


comp <- compare_smooths(ccf.model, ccf.nospp, smooths = 's(ccf.pl)')
draw(comp)

#most of what ccf is accounting for is density. There is a slight difference in these two curves, but I'm fairly confident that if ccf is accounting for species mix, that its relatively un-impactful. 

my.rmse.2(ccf.model)
my.rmse.2(ccf.nospp)
```

```{r}
#1 bai~dbh + bal + CR + BA + Asp.
re.tree.1alt

#2 bai~dbh + bal + CR + ccf + Asp.

ccf_noslope
  #gam(bai~s(DIAMETER) + s(cr, k = 9) + s(bal.pl.ratio) + s(ccf.pl) + s(asp_sin, asp_cos) + s(unique.tree.f, bs = 're'), family = 'Gamma'(link = log), data = bai.spmx.ids, method = 'ML', control = list(nthreads = 3))
# saveRDS(mod_2, 'data/model_objects.1/ccf_noslope.rds')
# saveRDS(mod_2, 'C:/Users/cm165878/Box/research_/ccf_noslope.rds')

#3 bai~dbh + bal + CR + (BA * %larch) + Asp.
spmx.gam.alt.lf #this is not an interaction version, this just has ba and larch% separately

#4 bai~dbh + bal + CR + (BA * shade) + Asp.
spmx.gam.alt.st #this does not include interaction, just ba and shade separately

#4 bai~dbh + bal + CR + (ccf * shade)??? + Asp.

```

```{r}
#how much does species mixing matter?
models <- c('ba_model', 'ccf_model', 'larch%', 'shade')
rmses <- c(my.rmse.2(re.tree.1alt)[[2]],
  my.rmse.2(ccf_noslope)[[2]],
  my.rmse.2(spmx.gam.alt.lf)[[2]],
  my.rmse.2(spmx.gam.alt.st)[[2]])*6.452
dev.exp <- 100*c(a$dev.expl, aa$dev.expl, b$dev.expl, c$dev.expl)
aics <- c(AIC(re.tree.1alt), AIC(ccf_noslope), AIC(spmx.gam.alt.lf), AIC(spmx.gam.alt.st))


tibble(models, rmses, dev.exp, aics)

lf.sm <- smooth_estimates(spmx.gam.alt.lf, smooth = 's(larch.ba.fraction.pl)') %>% add_confint()
st.sm <- smooth_estimates(spmx.gam.alt.st, 's(shade.tol.pl)') %>% add_confint()

ggplot(lf.sm, aes(x = larch.ba.fraction.pl, y = est)) + 
  geom_line() + geom_ribbon(aes(ymin = est-se, ymax = est+se), alpha = 0.2) + geom_hline(yintercept = 0, linetype = 'dashed') + 
  geom_line(data = st.sm, aes(x = shade.tol.pl, y = est), color = 'blue') + 
  geom_ribbon(inherit.aes = F, data = st.sm, aes(ymin = est-se, ymax = est+se, x = shade.tol.pl), alpha = 0.2, fill = 'blue') + 
  geom_rug(inherit.aes = F, data = bai.spmx.ids, aes(x = larch.ba.fraction.pl),alpha = 0.2) +
  geom_rug(inherit.aes = F, data = bai.spmx.ids, aes(x = shade.tol.pl), alpha = 0.2, color = 'blue', sides = 't') + labs(x = 'larch fraction (black) & shade-intolerance (blue)')

```

```{r}
#original models
spmx.re.gam.lf <- readRDS('data/model_objects.1/spmx.re.gam.lf.rds')
spmx.re.gam.st <- readRDS('data/model_objects.1/spmx.re.gam.st.rds')
lf.smest <- smooth_estimates(spmx.re.gam.lf, smooth = 's(larch.ba.fraction.pl)') %>% add_confint()

st.smest <- smooth_estimates(spmx.re.gam.st, smooth = 's(shade.tol.pl)') %>% add_confint()

f1a <- ggplot(lf.smest, aes(x = larch.ba.fraction.pl, y = est)) + 
  geom_ribbon(alpha = 0.2, aes(ymin = lower_ci, ymax = upper_ci, x = larch.ba.fraction.pl)) + 
  geom_line() +
  labs(x = 'Larch proportion') +
  scale_y_continuous(breaks = seq(-2, 8, by = 2)) + geom_hline(yintercept = 0, linetype = 'dashed')

f1b <- ggplot(st.smest, aes(x = shade.tol.pl, y = est)) +
  geom_ribbon(alpha = 0.2, aes(ymin = lower_ci, ymax = upper_ci, x = shade.tol.pl)) + 
  geom_line() +
  labs(x = 'Shade intolerance') +
  scale_y_continuous(breaks = seq(-2, 8, by = 2)) +
  scale_x_continuous(n.breaks = 5) + geom_hline(yintercept = 0, linetype = 'dashed')
f1a
f1b

ggplot() + geom_line(data = lf.smest, aes(x = larch.ba.fraction.pl, y = est)) + geom_ribbon(data = lf.smest, aes(x= larch.ba.fraction.pl, y = est, ymin = est-se, ymax = est+se), alpha = 0.2) + 
  geom_line(data = st.smest, aes(x = shade.tol.pl, y = est), color = 'blue') + geom_ribbon(data = st.smest, aes(x = shade.tol.pl, ymin = est-se, ymax = est+se), alpha = 0.2, fill = 'blue') +
  geom_hline(yintercept = 0, linetype = 'dashed') + geom_rug(inherit.aes = F, data = bai.spmx.ids, aes(x = larch.ba.fraction.pl),alpha = 0.2) +
  geom_rug(inherit.aes = F, data = bai.spmx.ids, aes(x = shade.tol.pl), alpha = 0.2, color = 'blue', sides = 't') + labs(x = 'larch fraction (black) & shade-intolerance (blue)')
```

```{r}

# spmx.gam.alt.st

#lf.sm <- smooth_estimates(spmx.gam.alt.lf, smooth = 's(larch.ba.fraction.pl)') %>% add_confint()
st.sm <- smooth_estimates(spmx.gam.alt.st, 's(shade.tol.pl)') %>% add_confint()

# create rug plot where mixture is categorized. selecting species that has >= 20% portion of basal area
# 
spp.cats <- bai.spmx.ids %>% 
  mutate(spp.cat.value = pmax(percent_PSME, percent_PIEN, 
                              percent_ABLA, percent_ABGR, 
                              percent_PICO, percent_PIPO, 
                              percent_other)) %>% 
  mutate(spp.cat = 
           case_when(spp.cat.value < .2 ~ 'laoc',
                     spp.cat.value == percent_PSME & spp.cat.value >= .2 ~ 'psme',
                     spp.cat.value == percent_PIEN & spp.cat.value >= .2 ~ 'pien',
                     spp.cat.value == percent_ABLA & spp.cat.value >= .2 ~ 'abla',
                     spp.cat.value == percent_ABGR & spp.cat.value >= .2 ~ 'abgr',
                     spp.cat.value == percent_PICO & spp.cat.value >= .2 ~ 'pico',
                     spp.cat.value == percent_PIPO & spp.cat.value >= .2 ~ 'pipo',
                     spp.cat.value == percent_other & spp.cat.value >= .2 ~ 'other'))

psme <- spp.cats %>% filter(spp.cat == 'psme')
pien <- spp.cats %>% filter(spp.cat == 'pien')
abla <- spp.cats %>% filter(spp.cat == 'abla')
abgr <- spp.cats %>% filter(spp.cat == 'abgr')
pico <- spp.cats %>% filter(spp.cat == 'pico')
pipo <- spp.cats %>% filter(spp.cat == 'pipo')
other <- spp.cats %>% filter(spp.cat == 'other')
laoc <- spp.cats %>% filter(spp.cat == 'laoc')

ggplot() + 
  geom_line(data = st.sm, aes(x = shade.tol.pl, y = est)) + 
  geom_ribbon(data = st.sm, aes(x = shade.tol.pl, ymin = est-se, ymax = est+se), alpha = 0.2) + 
  geom_hline(yintercept = 0, linetype = 'dashed') +
  geom_rug(inherit.aes = F, data = spp.cats, aes(x = shade.tol.pl, color = factor(spp.cat))) #+ theme(legend.position = 'none')

plot.psme <- ggplot(st.sm, aes(x = shade.tol.pl, y = est)) + 
  geom_line() + 
  geom_ribbon(aes(ymin = est-se, ymax = est+se), alpha = 0.2) + 
  geom_hline(yintercept = 0, linetype = 'dashed') +
  geom_rug(inherit.aes = F, data = psme, aes(x = shade.tol.pl)) + labs(title = 'psme')

plot.pien <- ggplot(st.sm, aes(x = shade.tol.pl, y = est)) + 
  geom_line() + 
  geom_ribbon(aes(ymin = est-se, ymax = est+se), alpha = 0.2) + 
  geom_hline(yintercept = 0, linetype = 'dashed') +
  geom_rug(inherit.aes = F, data = pien, aes(x = shade.tol.pl)) + labs(title = 'pien')

plot.abla <- ggplot(st.sm, aes(x = shade.tol.pl, y = est)) + 
  geom_line() + 
  geom_ribbon(aes(ymin = est-se, ymax = est+se), alpha = 0.2) + 
  geom_hline(yintercept = 0, linetype = 'dashed') +
  geom_rug(inherit.aes = F, data = abla, aes(x = shade.tol.pl)) + labs(title = 'abla')

plot.abgr <- ggplot(st.sm, aes(x = shade.tol.pl, y = est)) + 
  geom_line() + 
  geom_ribbon(aes(ymin = est-se, ymax = est+se), alpha = 0.2) + 
  geom_hline(yintercept = 0, linetype = 'dashed') +
  geom_rug(inherit.aes = F, data = abgr, aes(x = shade.tol.pl)) + labs(title = 'abgr')

plot.pico <- ggplot(st.sm, aes(x = shade.tol.pl, y = est)) + 
  geom_line() + 
  geom_ribbon(aes(ymin = est-se, ymax = est+se), alpha = 0.2) + 
  geom_hline(yintercept = 0, linetype = 'dashed') +
  geom_rug(inherit.aes = F, data = pico, aes(x = shade.tol.pl)) + labs(title = 'pico')

plot.pipo <- ggplot(st.sm, aes(x = shade.tol.pl, y = est)) + 
  geom_line() + 
  geom_ribbon(aes(ymin = est-se, ymax = est+se), alpha = 0.2) + 
  geom_hline(yintercept = 0, linetype = 'dashed') +
  geom_rug(inherit.aes = F, data = pipo, aes(x = shade.tol.pl)) + labs(title = 'pipo')

plot.oth <- ggplot(st.sm, aes(x = shade.tol.pl, y = est)) + 
  geom_line() + 
  geom_ribbon(aes(ymin = est-se, ymax = est+se), alpha = 0.2) + 
  geom_hline(yintercept = 0, linetype = 'dashed') +
  geom_rug(inherit.aes = F, data = other, aes(x = shade.tol.pl)) + labs(title = 'other')

plot.laoc <- ggplot(st.sm, aes(x = shade.tol.pl, y = est)) + 
  geom_line() + 
  geom_ribbon(aes(ymin = est-se, ymax = est+se), alpha = 0.2) + 
  geom_hline(yintercept = 0, linetype = 'dashed') +
  geom_rug(inherit.aes = F, data = laoc, aes(x = shade.tol.pl)) + labs(title = 'laoc')

grid.arrange(plot.laoc, plot.pico, plot.pipo, plot.psme, plot.abla, plot.pien, plot.oth)
grid.arrange(plot.abla, plot.pien, nrow=2)
plot.psme

```

10. here, I created figures of shade tol partial response with rug plots for each species that has >20% representation by basal area
 --- at the moment I see PICO and LAOC influence the curve at high shade-intolerance values (as expected). 
 --- there is a strong dip in the partial response curve that may be associated with a high presence pico and/or larch
 --- pipo values also appear to be present as the partial response curve dips at the higher-end of shade-intolerance values
 --- psme may be expressing a positive influence on growth. There are numerous psme data points associated with the peak around 0.75
 --- along the same lines, the peak at .75 may be influenced by the presence of pien, as that's where the most pien values seem to sit. 
 --- abla values are most present underneath the far left peak and a handful are present at the second peak (~0.75)

```{r}
#lf.sm <- smooth_estimates(spmx.gam.alt.lf, smooth = 's(larch.ba.fraction.pl)') %>% add_confint()
lf.sm <- smooth_estimates(spmx.gam.alt.lf, 's(percent_LAOC)') %>% add_confint()

# create rug plot where mixture is categorized.
# 
spp.cats <- bai.spmx.ids %>% 
  mutate(spp.cat.value = pmax(percent_PSME, percent_PIEN, 
                              percent_ABLA, percent_ABGR, 
                              percent_PICO, percent_PIPO, 
                              percent_other)) %>% 
  mutate(spp.cat = 
           case_when(spp.cat.value < .2 ~ 'laoc',
                     spp.cat.value == percent_PSME & spp.cat.value >= .2 ~ 'psme',
                     spp.cat.value == percent_PIEN & spp.cat.value >= .2 ~ 'pien',
                     spp.cat.value == percent_ABLA & spp.cat.value >= .2 ~ 'abla',
                     spp.cat.value == percent_ABGR & spp.cat.value >= .2 ~ 'abgr',
                     spp.cat.value == percent_PICO & spp.cat.value >= .2 ~ 'pico',
                     spp.cat.value == percent_PIPO & spp.cat.value >= .2 ~ 'pipo',
                     spp.cat.value == percent_other & spp.cat.value >= .2 ~ 'other'))

psme <- spp.cats %>% filter(spp.cat == 'psme')
pien <- spp.cats %>% filter(spp.cat == 'pien')
abla <- spp.cats %>% filter(spp.cat == 'abla')
abgr <- spp.cats %>% filter(spp.cat == 'abgr')
pico <- spp.cats %>% filter(spp.cat == 'pico')
pipo <- spp.cats %>% filter(spp.cat == 'pipo')
other <- spp.cats %>% filter(spp.cat == 'other')
laoc <- spp.cats %>% filter(spp.cat == 'laoc')

ggplot() + 
  geom_line(data = lf.sm, aes(x = percent_LAOC, y = est)) + 
  geom_ribbon(data = lf.sm, aes(x = percent_LAOC, ymin = est-se, ymax = est+se), alpha = 0.2) + 
  geom_hline(yintercept = 0, linetype = 'dashed') +
  geom_rug(inherit.aes = F, data = spp.cats, aes(x = percent_LAOC, color = factor(spp.cat))) #+ theme(legend.position = 'none')

plot.psme.lf <- ggplot(lf.sm, aes(x = percent_LAOC, y = est)) + 
  geom_line() + 
  geom_ribbon(aes(ymin = est-se, ymax = est+se), alpha = 0.2) + 
  geom_hline(yintercept = 0, linetype = 'dashed') +
  geom_rug(inherit.aes = F, data = psme, aes(x = percent_LAOC)) + labs(title = 'psme')

plot.pien.lf <- ggplot(lf.sm, aes(x = percent_LAOC, y = est)) + 
  geom_line() + 
  geom_ribbon(aes(ymin = est-se, ymax = est+se), alpha = 0.2) + 
  geom_hline(yintercept = 0, linetype = 'dashed') +
  geom_rug(inherit.aes = F, data = pien, aes(x = percent_LAOC)) + labs(title = 'pien')

plot.abla.lf <- ggplot(lf.sm, aes(x = percent_LAOC, y = est)) + 
  geom_line() + 
  geom_ribbon(aes(ymin = est-se, ymax = est+se), alpha = 0.2) + 
  geom_hline(yintercept = 0, linetype = 'dashed') +
  geom_rug(inherit.aes = F, data = abla, aes(x = percent_LAOC)) + labs(title = 'abla')

plot.abgr.lf <- ggplot(lf.sm, aes(x = percent_LAOC, y = est)) + 
  geom_line() + 
  geom_ribbon(aes(ymin = est-se, ymax = est+se), alpha = 0.2) + 
  geom_hline(yintercept = 0, linetype = 'dashed') +
  geom_rug(inherit.aes = F, data = abgr, aes(x = percent_LAOC)) + labs(title = 'abgr')

plot.pico.lf <- ggplot(lf.sm, aes(x = percent_LAOC, y = est)) + 
  geom_line() + 
  geom_ribbon(aes(ymin = est-se, ymax = est+se), alpha = 0.2) + 
  geom_hline(yintercept = 0, linetype = 'dashed') +
  geom_rug(inherit.aes = F, data = pico, aes(x = percent_LAOC)) + labs(title = 'pico')

plot.pipo.lf <- ggplot(lf.sm, aes(x = percent_LAOC, y = est)) + 
  geom_line() + 
  geom_ribbon(aes(ymin = est-se, ymax = est+se), alpha = 0.2) + 
  geom_hline(yintercept = 0, linetype = 'dashed') +
  geom_rug(inherit.aes = F, data = pipo, aes(x = percent_LAOC)) + labs(title = 'pipo')

plot.oth.lf <- ggplot(lf.sm, aes(x = percent_LAOC, y = est)) + 
  geom_line() + 
  geom_ribbon(aes(ymin = est-se, ymax = est+se), alpha = 0.2) + 
  geom_hline(yintercept = 0, linetype = 'dashed') +
  geom_rug(inherit.aes = F, data = other, aes(x = percent_LAOC)) + labs(title = 'other')

plot.laoc.lf <- ggplot(lf.sm, aes(x = percent_LAOC, y = est)) + 
  geom_line() + 
  geom_ribbon(aes(ymin = est-se, ymax = est+se), alpha = 0.2) + 
  geom_hline(yintercept = 0, linetype = 'dashed') +
  geom_rug(inherit.aes = F, data = laoc, aes(x = percent_LAOC)) + labs(title = 'laoc')

grid.arrange(plot.laoc.lf, plot.pico.lf, plot.pipo.lf, plot.psme.lf, plot.abla.lf, plot.pien.lf, plot.oth.lf)
```


```{r}
# spp.cats
laoc.sc <- ggplot(subset(spp.cats, percent_LAOC>.01), aes(x = percent_LAOC, shade.tol.pl)) + geom_point() + labs(title = 'LAOC') + scale_y_continuous(breaks = c(0,.25,.5, .75,1), limits = c(0,1))

psme.sc <- ggplot(subset(spp.cats, percent_PSME>.01), aes(x = percent_PSME, shade.tol.pl)) + geom_point() + labs(title = 'psme') + scale_y_continuous(breaks = c(0,.25,.5, .75,1), limits = c(0,1))

abla.sc <- ggplot(subset(spp.cats, percent_ABLA>.01), aes(x = percent_ABLA, shade.tol.pl)) + geom_point() + labs(title = 'abla') + scale_y_continuous(breaks = c(0,.25,.5, .75,1), limits = c(0,1))

abgr.sc <- ggplot(subset(spp.cats, percent_ABGR>.01), aes(x = percent_ABGR, shade.tol.pl)) + geom_point() + labs(title = 'abgr') + scale_y_continuous(breaks = c(0,.25,.5, .75,1), limits = c(0,1))

pien.sc <- ggplot(subset(spp.cats, percent_PIEN>.01), aes(x = percent_PIEN, shade.tol.pl)) + geom_point() + labs(title = 'pien') + scale_y_continuous(breaks = c(0,.25,.5, .75,1), limits = c(0,1))

pico.sc <- ggplot(subset(spp.cats, percent_PICO>.01), aes(x = percent_PICO, shade.tol.pl)) + geom_point() + labs(title = 'pico') + scale_y_continuous(breaks = c(0,.25,.5, .75,1), limits = c(0,1))

pipo.sc <- ggplot(subset(spp.cats, percent_PIPO>.01), aes(x = percent_PIPO, shade.tol.pl)) + geom_point() + labs(title = 'pipo') + scale_y_continuous(breaks = c(0,.25,.5, .75,1), limits = c(0,1))

oth.sc <- ggplot(subset(spp.cats, percent_other>.01), aes(x = percent_other, shade.tol.pl)) + geom_point() + labs(title = 'other') + scale_y_continuous(breaks = c(0,.25,.5, .75,1), limits = c(0,1))

grid.arrange(laoc.sc, pico.sc, pipo.sc, psme.sc, abgr.sc, abla.sc, pien.sc, oth.sc)




```

I created scatterplots which show the percent of a species (laoc, pico, pipo, psme, abla, abgr, pien, other) on the x-axis and the shade-intolerance values on the y-axis.
 --- both larch and pico seem to approach the expected value of 1.
 --- pipo appears to approach 0.75
 --- psme seems to approach 0.5 (maybe a bit higher). 
 --- abla approahces 0.25
 --- abgr may be approaching a point of 0.45-0.5, but there may not be enough data to say yet. This is somewhat interesting.
 --- pien looks to be appraoching 0.25 as its abundance increases. 

```{r}
# metric cateogrizing larch with each other species
# larch.mixtures <- function(data, species, maximum, combined_min){
#   a <- data %>% filter(percent_LAOC < maximum & '{species}' < maximum & (percent_LAOC+'{species}') > combined_min)
#   return(a)
# }
# bai.spmx.ids %>% filter(percent_LAOC>0.4 & percent_PSME>0.4 & (percent_LAOC+percent_PSME)>0.6)

# larch.mixtures(bai.spmx.ids, percent_PSME, maximum = 0.8, combined_min =.4)

spp.fun <- function(spp_min, spp_maximum, combined_min){
  a <- bai.spmx.ids %>% filter(percent_LAOC<spp_maximum & percent_LAOC>spp_min, 
                               percent_PSME<spp_maximum & percent_PSME>spp_min, 
                               (percent_LAOC+percent_PSME)>combined_min)
  b <- bai.spmx.ids %>% filter(percent_LAOC<spp_maximum & percent_LAOC>spp_min, 
                               percent_ABLA<spp_maximum & percent_ABLA>spp_min, 
                               (percent_LAOC+percent_ABLA)>combined_min)
  c <- bai.spmx.ids %>% filter(percent_LAOC<spp_maximum & percent_LAOC>spp_min, 
                               percent_ABGR<spp_maximum & percent_ABGR>spp_min, 
                               (percent_LAOC+percent_ABGR)>combined_min)
  d <- bai.spmx.ids %>% filter(percent_LAOC<spp_maximum & percent_LAOC>spp_min, 
                               percent_PIEN<spp_maximum & percent_PIEN>spp_min, 
                               (percent_LAOC+percent_PIEN)>combined_min)
  e <- bai.spmx.ids %>% filter(percent_LAOC<spp_maximum & percent_LAOC>spp_min, 
                               percent_PICO<spp_maximum & percent_PICO>spp_min, 
                               (percent_LAOC+percent_PICO)>combined_min)
  f <- bai.spmx.ids %>% filter(percent_LAOC<spp_maximum & percent_LAOC>spp_min, 
                               percent_PIPO<spp_maximum & percent_PIPO>spp_min, 
                               (percent_LAOC+percent_PIPO)>combined_min)
  g <- bai.spmx.ids %>% filter(percent_LAOC>combined_min)
   return(list('psme'=a, 'abla'=b, 'abgr'=c, 'pien'=d, 'pico'=e, 'pipo'=f, 'laoc' = g))
  }
(mixtures <- spp.fun(0.2, 0.65, 0.7))
individual.obs <- lapply(mixtures, function(x)
  length(unique(x$unique_tree_id)))
(dbh.plots <- lapply(mixtures, function(m)
  p <- ggplot(m, aes(x = DIAMETER)) + geom_histogram()))
(dens.plots <- lapply(mixtures, function(m)
  p <- ggplot(m, aes(x = ba.pl)) + geom_histogram()))
(bai.plots <- lapply(mixtures, function(m)
  p <- ggplot(m, aes(x = bai)) + geom_histogram()))

data.frame(unlist(individual.obs))

dbh.plots


# bai.spmx.ids %>% filter(percent_LAOC<.6, percent_PSME<.6, (percent_LAOC+percent_PSME)>.6)
# bai.spmx.ids %>% filter(percent_LAOC<.6 & percent_ABLA<.6 & (percent_LAOC+percent_ABLA)>.8)
# bai.spmx.ids %>% filter(percent_LAOC<.6 & percent_ABGR<.6 & (percent_LAOC+percent_ABGR)>.8)
# bai.spmx.ids %>% filter(percent_LAOC<.6 & percent_PICO<.6 & (percent_LAOC+percent_PICO)>.8)
# bai.spmx.ids %>% filter(percent_LAOC<.6 & percent_PIPO<.6 & (percent_LAOC+percent_PIPO)>.8)
# bai.spmx.ids %>% filter(percent_LAOC<.6 & percent_PIEN<.6 & (percent_LAOC+percent_PIEN)>.8)



```

```{r}
#fit gams for subsets of data
laoc_psme <- mixtures$psme
hist(laoc_psme$bai)
length(unique(laoc_psme$asp_sin))
ggplot(laoc_psme, aes(x = percent_LAOC, y = percent_PSME)) + geom_point()
psme.gam <- gamm(bai~ s(DIAMETER) + s(cr, k = 8) + s(bal.pl.ratio) + s(ba.pl)+
                      s(asp_sin, asp_cos, k = 20),
                    random=list(unique_tree_id=~1),
                    family = 'Gamma'(link = log),
                    data = laoc_psme, method = 'ML')

laoc_pico <- mixtures$pico
length(unique(laoc_pico$cr))
hist(laoc_pico$bai)
plot(laoc_pico$DIAMETER, laoc_pico$bai)
filter(laoc_pico, bai>3)
ggplot(laoc_pico, aes(x = percent_LAOC, y = percent_PICO)) + geom_point()
pico.gam <- gamm(bai~ s(DIAMETER) + s(cr, k = 7) + s(bal.pl.ratio) + s(ba.pl)+
                      s(asp_sin, asp_cos, k = 13),
                    random=list(unique_tree_id=~1),
                    family = 'Gamma'(link = log),
                    data = laoc_pico, method = 'ML')

laoc_laoc <- mixtures$laoc
length(unique(laoc_laoc$asp_sin))
laoc.gam <- gamm(bai~ s(DIAMETER) + s(cr, k = 9) + s(bal.pl.ratio) + s(ba.pl)+
                      s(asp_sin, asp_cos, k = 20),
                    random=list(unique_tree_id=~1),
                    family = 'Gamma'(link = log),
                    data = laoc_laoc, method = 'ML')

laoc_pien <- mixtures$pien
ggplot(laoc_pien, aes(x = percent_LAOC, y = percent_PIEN)) + geom_point()
length(unique(laoc_pien$asp_sin))
length(unique(laoc_pien$cr))
pien.gam <- gamm(bai~ s(DIAMETER) + s(cr, k = 8) + s(bal.pl.ratio) + s(ba.pl)+
                      s(asp_sin, asp_cos, k = 8),
                    random=list(unique_tree_id=~1),
                    family = 'Gamma'(link = log),
                    data = laoc_pien, method = 'ML')

mod.list <- list(psme.gam$gam, pico.gam$gam, pien.gam$gam, laoc.gam$gam)
lapply(mod.list, gam.check)
# plot(pico.gam$gam)
# filter(bai.spmx.ids, unique_tree_id == 2505)
# bbb <- readRDS('data/bai.shade.rds')
# bbb %>% filter(unique_tree_id == 2505)
# bbb %>% filter(!(unique_tree_id == 2505 & MEASUREMENT_NO == 4))
# rm(data)

## change bai to 4.611038134
plot(psme.gam$gam, residuals = T, pages = 1)
plot(pico.gam$gam, residuals = T, pages = 1)
plot(laoc.gam$gam, residuals = T, pages = 1)
plot(pien.gam$gam, residuals = T, pages = 1)
```

```{r}
# max(as.numeric(bai.spmx.ids$unique_tree_id))
#simulated data
DIAMETER <- (seq(1, 10, length = 100))
# cr <- sample(seq(.5,.9,0.1), 100, replace = T)
# bal.pl.ratio <- sample(seq(0.3658, 0.8419), 100, replace = T)
# ba.pl <- sample(seq(25, 200, 1), 100, replace = T)
asp_sin <- rep(sin(median(bai.spmx.ids$aspect_deg)*(pi/180)), 100)
asp_cos <- rep(cos(median(bai.spmx.ids$aspect_deg)*(pi/180)), 100)

#varying diameter values - everything else constant

cr <- rep(mean(bai.spmx.ids$cr, na.rm = T), 100)
bal.pl.ratio <- rep(mean(bai.spmx.ids$bal.pl.ratio, na.rm = T), 100)
ba.pl <- rep(mean(bai.spmx.ids$ba.pl, na.rm = T), 100)

# hist(laoc_laoc$bai)

var.diam <- tibble(DIAMETER, cr, bal.pl.ratio, ba.pl, asp_sin, asp_cos)

bai.diam.a <- predict(laoc.gam$gam, newdata = var.diam, type = 'response')
a <- tibble(var.diam, bai = as.vector(bai.diam.a))

bai.diam.b <- predict(psme.gam$gam, newdata = var.diam, type = 'response')
b <- tibble(var.diam, bai = as.vector(bai.diam.b))

bai.diam.c <- predict(pico.gam$gam, newdata = var.diam, type = 'response')
c <- tibble(var.diam, bai = as.vector(bai.diam.c))

rm(DIAMETER)
ggplot() + geom_point(data = a, aes(x = DIAMETER, y = bai), color = 'black') + 
  geom_point(data = b, aes(x = DIAMETER, y = bai), color = 'blue') + 
  geom_point(data = c, aes(x = DIAMETER, y = bai), color = 'orange')

DIAMETER <- rep(median(bai.spmx.ids$DIAMETER), 100)
ba.pl <- seq(min(bai.spmx.ids$ba.pl), max(bai.spmx.ids$ba.pl), length = 100)
var.ba <- tibble(DIAMETER, cr, bal.pl.ratio, ba.pl, asp_sin, asp_cos)

bai.ba.a <- predict(laoc.gam$gam, newdata = var.ba, type = 'response')
aa <- tibble(var.ba, bai = as.vector(bai.ba.a))

bai.ba.b <- predict(psme.gam$gam, newdata = var.ba, type = 'response')
bb <- tibble(var.ba, bai = as.vector(bai.ba.b))

bai.ba.c <- predict(pico.gam$gam, newdata = var.ba, type = 'response')
cc <- tibble(var.ba, bai = as.vector(bai.ba.c))

rm(ba.pl)
ggplot() + 
  geom_point(data = aa, aes(x = ba.pl, y = bai), color = 'black') +
  geom_point(data = bb, aes(x = ba.pl, y = bai), color = 'blue') #+
  #geom_point(data = cc, aes(x = ba.pl, y = bai), color = 'orange')

```
