==> 13.c.gam.fit.site.v2.Rmd <==
---
title: "gam.fit.site.v2"
author: "Christian Mercado"
date: "4/19/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE)

#load packages
library(tidyverse)
library(mgcv)
library(gratia)
library(ggrepel)
library(corrplot)
library(GGally)
# library(mgcViz)
source('12.gam_fns.R')

hab_loc_codes <- read_csv('data/habtypes.csv')
hab_loc_codes <- hab_loc_codes %>% select(SETTING_ID, habclass, locationcode)

#load data
bai.train <- readRDS('data/bai.train.7_14_22.rds')
bai.train <- left_join(bai.train, hab_loc_codes, by = "SETTING_ID")

# bai.size <- bai.train %>% 
#   select(stand, MEASUREMENT_NO, PLOT, unique_tree_id, 
#          bai, DIAMETER, treeba, log.diam, log.bai, grwth.yr) %>% 
#   filter(bai != 0)
# 
# bai.site <- bai.train %>% 
#   select(stand, MEASUREMENT_NO, PLOT, unique_tree_id, 
#          bai, DIAMETER, treeba, log.diam, log.bai, grwth.yr, mean_si, aspect_deg, 
#          heatload, slope_deg, elev_m, NF, CROWN_RATIO, tpa.pl.all, tpa.pl.cutoff,
#          ba.pl, bal.pl, ccf.pl, qmd.pl.all, qmd.pl.cutoff, dq.pl.all, dq.pl.cutoff, tpa.cl.all, tpa.pl.cutoff,
#          ba.cl, ccf.cl, bal.cl, dq.cl.all, dq.cl.cutoff, HabType, habclass,
#          locationcode) %>% 
#   mutate(slope_pct = tan(slope_deg*(pi/180))*100, 
#          asp_sin = sin(aspect_deg*(pi/180)),
#          asp_cos = cos(aspect_deg*(pi/180))) %>%
#   mutate(sl_asp_sin = asp_sin*slope_pct, 
#          sl_asp_cos = asp_cos*slope_pct, 
#          asp.trasp = trasp(aspect_deg), 
#          HabType = as.factor(HabType),
#          cr = CROWN_RATIO/100,
#          bal.pl.ratio = bal.pl/ba.pl) %>% 
#   filter(bai != 0)

bai.spmx <- bai.train %>% 
  select(SETTING_ID, stand, MEASUREMENT_NO, myear, PLOT, cluster, unique_tree_id, 
         bai, DIAMETER, treeba, log.diam, log.bai, grwth.yr, mean_si, aspect_deg, 
         heatload, slope_deg, elev_m, NF, CROWN_RATIO, tpa.pl.all, tpa.pl.cutoff,
         ba.pl, bal.pl, ccf.pl, ccf.nospp, qmd.pl.all, qmd.pl.cutoff, dq.pl.all, 
         dq.pl.cutoff, tpa.cl.all, tpa.pl.cutoff,
         ba.cl, ccf.cl, bal.cl, dq.cl.all, dq.cl.cutoff, HabType, habclass,
         locationcode, percent_LAOC, shade.tol.pl, dom.spp.pl.ba, percent_PSME, 
         percent_PIEN, percent_ABLA, percent_ABGR, percent_PICO, percent_PIPO, percent_other) %>% 
  mutate(slope_pct = tan(slope_deg*(pi/180))*100, 
         asp_sin = sin(aspect_deg*(pi/180)),
         asp_cos = cos(aspect_deg*(pi/180))) %>%
  mutate(sl_asp_sin = asp_sin*slope_pct, 
         sl_asp_cos = asp_cos*slope_pct, 
         asp.trasp = trasp(aspect_deg), 
         HabType = as.factor(HabType),
         cr = CROWN_RATIO/100,
         bal.pl.ratio = bal.pl/ba.pl,
         unique_tree_id = factor(unique_tree_id)) %>% 
  filter(bai != 0)


bai.spmx.ids <- bai.spmx %>% 
  group_by(stand) %>% 
  mutate(unique.stand = group_indices()) %>%
  ungroup() %>% 
  mutate(unique.cluster = (1000*unique.stand)+cluster, 
         unique.plot = (1000*unique.stand)+PLOT) %>% 
  mutate(unique.cluster.meas = as.factor(unique.cluster + MEASUREMENT_NO), 
         unique.plot.meas = as.factor(unique.plot + MEASUREMENT_NO),
         unique.cluster.f = as.factor(unique.cluster), 
         unique.plot.f = as.factor(unique.plot), 
         unique.tree.f = as.factor(unique_tree_id)) %>% 
  filter(unique_tree_id != 15431)

gam.s1a <- gam(bai~s(DIAMETER), family = 'Gamma'(link = log), data = bai.spmx.ids, method = 'ML') #without giant trees
```

```{r winners from compdens}
## CHECK HERE - have these changed?
op2 <- gam(bai~s(DIAMETER) + s(cr, k = 9) + s(bal.pl.ratio) + s(dq.pl.all) + s(qmd.pl.all) + s(ccf.pl), family = 'Gamma'(link = log), data = bai.spmx.ids, method = 'ML')

compd.pl4 <- gam(bai~s(DIAMETER) + s(cr, k = 9) + s(bal.pl.ratio) + s(ccf.pl), family = 'Gamma'(link = log), data = bai.spmx.ids, method = 'ML')

rmse.conc(op2)[[1]]
rmse.conc(compd.pl4)[[1]]
```


```{r corrs}

#slope, aspect, elev
topo <- bai.spmx.ids %>% select(slope_pct, asp_sin, asp_cos, elev_m)
#site index
#heatload
#setting id
allsite <- bai.spmx.ids %>% select(slope_pct, asp_sin, asp_cos, elev_m, mean_si, heatload)

ggpairs(topo)
ggpairs(allsite)

```

```{r 1 old version, include = F}

site1 <- update(op2, . ~ . + s(slope_pct))

site2 <- update(op2, . ~ . + s(elev_m))

site3 <- update(op2, . ~ . + s(asp_sin))

site4 <- update(op2, . ~ . + s(asp_cos))

site5 <- update(op2, . ~ . + s(asp_sin, asp_cos))

site6 <- update(op2, . ~ . + s(asp_sin) + s(asp_cos) + ti(asp_sin, asp_cos))

sitelist1 <- list('slope' = site1, 'elev' = site2, 'sinasp' = site3, 'cosasp' = site4, 'sincosasp' = site5, 'asp_decomposed' = site6)

rmse.conc.table(sitelist1)

## sincos wins barely
## too high concurvity - elevation is next

#elevation options
site2a <- update(site2, . ~ . + s(slope_pct))
site2b <- update(site2, . ~ . + s(asp_sin, asp_cos))
site2c <- update(site2, . ~ . + s(asp_sin))
site2d <- update(site2, . ~ . + s(asp_cos))

sitelist2 <- list('el-sl' = site2a, 'el-sincos' = site2b, 'el-sin' = site2c, 'el-cos' = site2d)
#lapply(sitelist2, rmse.conc)
rmse.conc.table(sitelist2)

#sincos versions
site5a <- update(site5, . ~ . + s(slope_pct))
site5b <- update(site5, . ~ . + s(elev_m))
site5c <- update(site5, . ~ . + s(slope_pct) + s(elev_m))

sitelist3 <- list('sincos-sl' = site5a, 'sincos-el' = site5b, 'sincos-sl-el' = site5c)
#lapply(sitelist2, rmse.conc)
rmse.conc.table(sitelist3)
#el-sl is the technical winner here

```

```{r 2 old - evaluate selections, include=FALSE}
summary(site2)
draw(site2, rug = F, residuals = T, scales = 'free')
par(mfrow = c(2,2))
gam.check(site2)

#look at best model
summary(site2a)
draw(site2a, rug = F, residuals = T, scales = 'free')

#appraise(site2a)
par(mfrow = c(2,2))
gam.check(site2a)
```

```{r 1 new version choose best}
site.n1a <- update(compd.pl4, . ~ . + s(slope_pct))

site.n2a <- update(compd.pl4, . ~ . + s(elev_m))

site.n3a <- update(compd.pl4, . ~ . + s(asp_sin))

site.n4a <- update(compd.pl4, . ~ . + s(asp_cos))

site.n5a <- update(compd.pl4, . ~ . + s(asp_sin, asp_cos))

site.n6a <- update(compd.pl4, . ~ . + s(asp_sin) + s(asp_cos) + ti(asp_sin, asp_cos))

sitelist1a <- list('slope' = site.n1a, 'elev' = site.n2a, 'sinasp' = site.n3a, 'cosasp' = site.n4a, 'sincosasp' = site.n5a, 'asp_decomposed' = site.n6a)

rmse.conc(compd.pl4)[[1]]
rmse.conc.table(sitelist1a)
#is the change in rmse provided by variable at least half of that of the variable which changes it the most?
rmse.conc.table(sitelist1a) %>% mutate(rmse.change = rmse.conc(compd.pl4)[[1]]-rmse) %>% mutate(rmse.change.lg = rmse.change>=(.5*rmse.change[rmse.rank == 1]))


## in this case, then, slope, elev, and asp_decomposed all have a difference of at least half of that of sincos, the 'best' variable by this method. note that concurvity for the decomposed sin-cos version is high (due to the inherent relationship among these)

## ranks then are
## sincos, elv, slope
```

```{r 1 ALT - BA version}
compd.ba1 <- gam(bai~s(DIAMETER) + s(cr, k = 9) + s(bal.pl.ratio) + s(ba.pl), family = 'Gamma'(link = log), data = bai.spmx.ids, method = 'ML')

site.n1b <- update(compd.ba1, . ~ . + s(slope_pct))

site.n2b <- update(compd.ba1, . ~ . + s(elev_m))

site.n3b <- update(compd.ba1, . ~ . + s(asp_sin))

site.n4b <- update(compd.ba1, . ~ . + s(asp_cos))

site.n5b <- update(compd.ba1, . ~ . + s(asp_sin, asp_cos))

site.n6b <- update(compd.ba1, . ~ . + s(asp_sin) + s(asp_cos) + ti(asp_sin, asp_cos))

sitelist1b <- list('slope' = site.n1b, 'elev' = site.n2b, 'sinasp' = site.n3b, 'cosasp' = site.n4b, 'sincosasp' = site.n5b, 'asp_decomposed' = site.n6b)

rmse.conc.table(sitelist1b)

#lapply(sitelist1b, rmse.conc)
rmse.conc(compd.ba1)
rmse.conc(site.n5b)
```


```{r 1a summaries and diagnostics, include = F}
#sincos
summary(site.n5a)
draw(site5a, rug = F, residuals = T, scales = 'free')
par(mfrow = c(2,2))
gam.check(site.n5a)

#elev
summary(site.n2a)
draw(site2a, rug = F, residuals = T, scales = 'free')
par(mfrow = c(2,2))
gam.check(site.n2a)

#slope
summary(site.n1a)
draw(site.n1a, rug = F, residuals = T, scales = 'free')
par(mfrow = c(2,2))
gam.check(site.n1a)
```

```{r compare smooths, include = F}
for(model in sitelist1a){
  for(i in 1:length(model$var.summary)){
    par(mfrow = c(1,2))
    plot.gam(compd.pl4, rug = F, residuals = T, ylim = c(-3,4), select = i, scheme = 1)
    plot.gam(model, rug = F, residuals = T, ylim = c(-3,4), select = i, scheme = 1)  
    }
}

```


```{r 2 new version}
site5a.1 <- update(site.n5a, . ~ . + s(elev_m))
site5a.2 <- update(site.n5a, . ~ . + s(slope_pct))

lst.5 <- list('sincos-el' = site5a.1, 'sincos-slp' = site5a.2)

rmse.conc.table(lst.5)

site5a.3 <- update(site5a.1, . ~ . + s(slope_pct))

lapply(lst.5, rmse.conc)
rmse.conc(site5a.3)


## best is site5a.2 - sincos + slope. Adding elevation (site5a.3) also reduces RMSE and keeps concurvity low enough, except concurvity definitely jumped when slope and elev are together.. 

#for now going with site5a.2 - sincos-slope
## although technically site5a.3 is the best by RMSE

```

```{r 2 - ALT - ba version}


site5b.1 <- update(site.n5b, . ~ . + s(elev_m))
site5b.2 <- update(site.n5b, . ~ . + s(slope_pct))

rmse.conc(site5b.1)
rmse.conc(site5b.2)
```


```{r 2a new version}
#check summaries and diagnostics
summary(site5a.2)
plot.gam(site5a.2, rug = F, residuals = T, pages = 3, scheme = 1, ylim = c(-5,4))
par(mfrow = c(2,2))
gam.check(site5a.2)

```


```{r 3 new version}

#now compare raw topo vars with others
#topo
#site5a.2
#others
site.htld <- update(compd.pl4, . ~ . + s(heatload))
site.si <- update(compd.pl4, . ~ . + s(mean_si))
site.stand <- update(compd.pl4, . ~ . + stand)

site.lst <- list('sincos-sl' = site5a.2, 'site.heatload' = site.htld, 'site.si' = site.si, 'site.stand' = site.stand)

rmse.conc.table(site.lst)
lapply(site.lst, rmse.conc)


formula.gam(site5a.2)
## the topo variables win here - site5a.2
## the version using different stand intercepts is actually the winner by RMSE, but has too high concurvity (for intercepts as well as diameter)
```

```{r 3 - ALT - ba version}

#site.n5b

#site5b.1
#site5b.2

site.htld.ba <- update(compd.ba1, . ~ . + s(heatload))
site.si.ba <- update(compd.ba1, . ~ . + s(mean_si))
site.stand.ba <- update(compd.ba1, . ~ . + stand)

site.lst.ba <- list('sincos.ba' = site.n5b, 'heatload.ba' = site.htld.ba, 'si.ba' = site.si.ba, 'stand.ba' = site.stand.ba)

rmse.conc.table(site.lst.ba)

rmse.conc(site.stand.ba)

```


```{r FVS}
fvs.full <- gam(bai~elev_m + I(elev_m^2) + sl_asp_sin + sl_asp_cos + slope_pct + I(slope_pct^2) + I(ccf.cl/100) + log.diam + I(bal.cl/100) + cr + I(cr^2) + I(DIAMETER^2) + I(bal.cl/(100*log(DIAMETER+1))) + factor(habclass), data = bai.spmx.ids, family = 'Gamma'(link = log))

summary(fvs.full)


fvs.sz.cmp <- gam(bai~I(ccf.cl/100) + log.diam + I(bal.cl/100) + cr + I(cr^2) + I(DIAMETER^2) + I(bal.cl/(100*log(DIAMETER+1))) + factor(habclass), data = bai.spmx.ids, family = 'Gamma'(link = log))
summary(fvs.sz.cmp)

fvs.size1 <- gam(bai~log.diam + I(DIAMETER^2), family = 'Gamma'(link = log), data = bai.size[bai.size$unique_tree_id!=15431,])

```

```{r null model}
#create null model
gam.null <- gam(bai~1, family = 'Gamma'(link = log), data = bai.size)
summary.gam(gam.null)
my.rmse.2(gam.null)

best.6a <- gam(bai~s(DIAMETER) + s(ccf.pl) + s(cr, k = 9), family = 'Gamma'(link = log), bai.spmx.ids)
```

```{r RMSE table}

model <- c('null', 'size', 'density.comp', 'site')

rmse.selected <- c(my.rmse.2(gam.null)[[2]], my.rmse.2(gam.s1a)[[2]], my.rmse.2(compd.pl4)[[2]], my.rmse.2(site5a.2)[[2]])
#rmse.selected.old <- c(my.rmse.2(gam.null)[[2]], my.rmse.2(gam.s1a)[[2]], my.rmse.2(op2)[[2]], my.rmse.2(site2a)[[2]])
rmse.best1 <- c(my.rmse.2(gam.null)[[2]], my.rmse.2(gam.s1a)[[2]], my.rmse.2(best.6a)[[2]], my.rmse.2(site.n5a)[[2]])
fvs <- c(my.rmse.2(gam.null)[[2]], my.rmse.2(fvs.size1)[[2]], my.rmse.2(fvs.sz.cmp)[[2]], my.rmse.2(fvs.full)[[2]])

tibble(model, rmse.selected, rmse.best1, fvs) #rmse.selected.old

```
